Switch graph to period instead of frequency
This unit is easier to work with since we're reasoning about low numbers
of nanoseconds instead of hundreds of megahertz. The collate script was
modified to include both frequency and period measurements in its output.

Note that the y range was changed to a fixed range otherwise the
drastically higher exception periods obscure detail in other data.